requirements.txt
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.5
django-colorfield==0.1.15
django-debugtools==1.7.4
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-grappelli==2.13.1
django-import-export==1.2.0
django-menu-generator==1.0.4
dohq-tfs==1.0.172
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
MarkupPy==1.14
ntlm-auth==1.4.0
odfpy==1.4.0
openpyxl==3.0.1
pycparser==2.19
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-ntlm==1.1.0
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
urllib3==1.25.6
xlrd==1.2.0
xlwt==1.3.0
